# node-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Symptom Logger API",
    "version": "1.0.0",
    "description": "API documentation for the Symptom Logger API"
  },
  "servers": [
    {
      "url": "http://localhost:5200"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/admin/activity": {
      "get": {
        "summary": "Returns the number of unique users per day",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true,
            "description": "The start date for the activity log"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true,
            "description": "The end date for the activity log"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of unique users per day",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date"
                      },
                      "uniqueUsers": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/admin/geographic": {
      "get": {
        "summary": "Returns the symptom counts for the given location",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "country",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The country name"
          },
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The city name"
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "description": "The state name (optional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Symptom counts for the given location",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "symptom": {
                        "type": "string"
                      },
                      "count": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/admin/symptoms": {
      "get": {
        "summary": "Returns the patterns of the given symptom",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "symptom",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The symptom name"
          }
        ],
        "responses": {
          "200": {
            "description": "Symptom patterns for the given symptom",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "count": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/country": {
      "post": {
        "summary": "Create a new country",
        "tags": [
          "Country"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "alpha2": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2
                  },
                  "active": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Country created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "Get all countries",
        "tags": [
          "Country"
        ],
        "responses": {
          "200": {
            "description": "A list of countries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "alpha2": {
                        "type": "string"
                      },
                      "active": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/country/{id}": {
      "put": {
        "summary": "Update an existing country",
        "tags": [
          "Country"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The country ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "alpha2": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2
                  },
                  "active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Country updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Country not found"
          }
        }
      },
      "get": {
        "summary": "Get a country by ID",
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The country ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Country details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "alpha2": {
                      "type": "string"
                    },
                    "active": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Country not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a country by ID",
        "tags": [
          "Country"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The country ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Country deleted successfully"
          },
          "404": {
            "description": "Country not found"
          }
        }
      }
    },
...
  },
...
}